+++ Editor/PSDImportPostProcessor.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using UnityEditor.U2D.Sprites;
 using UnityEngine;
 
 namespace UnityEditor.Experimental.U2D.PSD