+++ Editor/SpriteData.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using UnityEditor.U2D.Sprites;
 using UnityEngine;
 using UnityEngine.Experimental.U2D;