+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.0-preview.3] - 2019-04-24
+## [1.2.0-preview.1] - 2019-03-15
 ### Added
-- Use latest 2D Animation package for 2019.1
+- Update support for 2019.2
+- Integrate with 2D Animation Sprite Library
+- Integrate with new 2D Animation Character Group
+- Fix asset name conflict
 
 ## [1.1.0-preview.2] - 2019-04-23
 ### Added